# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Patients
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients with optional search
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name, patientId, or phone
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  # OP Records
  /op-records:
    get:
      operationId: listOpRecords
      tags: [op-records]
      summary: List OP records
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of OP records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpRecord"
    post:
      operationId: createOpRecord
      tags: [op-records]
      summary: Create a new OP record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpRecord"
  /op-records/{id}:
    get:
      operationId: getOpRecord
      tags: [op-records]
      summary: Get OP record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OP Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpRecord"
    patch:
      operationId: updateOpRecord
      tags: [op-records]
      summary: Update OP record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpRecordUpdate"
      responses:
        "200":
          description: Updated OP record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpRecord"

  # Revisits
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: paymentStatus
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # Follow-ups
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role, status]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [administrator, receptionist, doctor]
        status:
          type: string
          enum: [active, inactive]
        lastLoginAt:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [username, password, role, fullName]
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [administrator, receptionist, doctor]
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
          enum: [administrator, receptionist, doctor]
        status:
          type: string
          enum: [active, inactive]
        password:
          type: string
    Patient:
      type: object
      required: [id, patientId, fullName, age, gender, contactNumber, status, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: string
        fullName:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [male, female, other]
        contactNumber:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        bloodGroup:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
    PatientInput:
      type: object
      required: [fullName, age, gender, contactNumber]
      properties:
        fullName:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [male, female, other]
        contactNumber:
          type: string
        email:
          type: string
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        bloodGroup:
          type: string
    PatientUpdate:
      type: object
      properties:
        fullName:
          type: string
        age:
          type: integer
        gender:
          type: string
        contactNumber:
          type: string
        email:
          type: string
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        bloodGroup:
          type: string
        status:
          type: string
          enum: [active, archived]
    OpRecord:
      type: object
      required: [id, opNumber, patientId, startDate, expiryDate, status, visitReason, createdAt]
      properties:
        id:
          type: integer
        opNumber:
          type: string
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        startDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [active, expired]
        visitReason:
          type: string
        createdAt:
          type: string
    OpRecordInput:
      type: object
      required: [patientId, visitReason]
      properties:
        patientId:
          type: integer
        visitReason:
          type: string
    OpRecordUpdate:
      type: object
      properties:
        visitReason:
          type: string
        status:
          type: string
          enum: [active, expired]
...
    Invoice:
      type: object
      required: [id, invoiceNumber, patientId, subtotal, total, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        opRecordId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        total:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid, partial]
        paymentMethod:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceItem:
      type: object
      required: [description, amount]
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        amount:
          type: number
    InvoiceInput:
      type: object
      required: [patientId, items, subtotal, total]
      properties:
        patientId:
          type: integer
        opRecordId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        total:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid, partial]
        paymentMethod:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        paymentStatus:
          type: string
          enum: [pending, paid, partial]
        paymentMethod:
          type: string
        paymentDate:
          type: string
...